requirements.txt
django>=3.1
django-environ
django-storages
google-cloud-storage
google-auth
Pillow
martor
psycopg2-binary
gunicorn
whitenoise
django-photologue
